+++ src/components/Body/Body.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import BrowseByCuisine from "./Cuisine-Section/BrowseByCuisine";
-import RestaurantList from "./Restaurant-Section/RestaurantList";
 import AllRestaurantDelivery from "./Restaurant-Section/AllRestaurantDelivery";
+import TopRestaurant from "./Restaurant-Section/TopRestaurant";
 
 function Body() {
     const [browseByCuisineData, setBrowseByCuisineData] = useState([]);
@@ -17,7 +17,7 @@ function Body() {
     // console.log(result?.data?.cards[1]?.card?.card?.gridElements?.infoWithStyle?.restaurants);
     setBrowseByCuisineData(result?.data?.cards[0]?.card?.card?.imageGridCards?.info);
     setRestaurantListData(result?.data?.cards[1]?.card?.card?.gridElements?.infoWithStyle?.restaurants);
-    setAllRestaurantListData(result?.data?.cards[1]?.card?.card?.gridElements?.infoWithStyle?.restaurants)
+    //setAllRestaurantListData(result?.data?.cards[1]?.card?.card?.gridElements?.infoWithStyle?.restaurants)
     console.log(result?.data);
 
   }
@@ -29,8 +29,8 @@ function Body() {
   return (
     <div className="w-[75%] mx-auto mt-3 overflow-hidden">
       <BrowseByCuisine  data={browseByCuisineData}/>
-      <RestaurantList data = {restaurantListData}/>
-      <AllRestaurantDelivery data={allRestaurantListData} />
+      <TopRestaurant data = {restaurantListData}/>
+      <AllRestaurantDelivery data={restaurantListData} />
 
     </div>
   );
